Fixed warnings

+++ StortSpelProjekt/BasicRenderer.cpp
@@ -26,7 +26,7 @@ bool BasicRenderer::setUpInputLayoutAnim(ID3D11Device* device, const std::string
 		{"WEIGHT", 0, DXGI_FORMAT_R32G32B32A32_FLOAT, 0, D3D11_APPEND_ALIGNED_ELEMENT, D3D11_INPUT_PER_VERTEX_DATA, 0}
 	};
 
-	HRESULT hr = device->CreateInputLayout(inputDesc, std::size(inputDesc), vShaderByteCode.c_str(), vShaderByteCode.length(), &this->animLayout);
+	HRESULT hr = device->CreateInputLayout(inputDesc, (UINT)std::size(inputDesc), vShaderByteCode.c_str(), vShaderByteCode.length(), &this->animLayout);
 	return !FAILED(hr);
 }
